requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
-f https://data.pyg.org/whl/torch-1.10.2+cu113.html
attrs==22.2.0
compress_pickle==2.1.0
igraph==0.9.11
karateclub==1.2.3
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
PyYAML==6.0
scikit_learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
tensorboard==2.10.1
tensorboardX==2.6
torch==1.10.2+cu113
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.64.1
